alters build target to 64bit and base SDK to 10.8, refactors to use appropriate types

+++ CCMenu/Classes/CCMServerMonitor.m
@@ -90,7 +90,7 @@ NSString *CCMProjectStatusUpdateNotification = @"CCMProjectStatusUpdateNotificat
         return;
     }
 	[NSTimer scheduledTimerWithTimeInterval:0.1 target:self selector:@selector(pollServers:) userInfo:nil repeats:NO];
-	int interval = [defaultsManager pollInterval];
+	NSInteger interval = [defaultsManager pollInterval];
 	timer = [NSTimer scheduledTimerWithTimeInterval:interval target:self selector:@selector(pollServers:) userInfo:nil repeats:YES];
 }
